refine converter codes
- Rename `LocalConverter` to `Converter`;
- Should enable `UpdateLayerDiffID` func ability in exported `Converter`;
- Move `LocalProvider` codes into separate file;
